+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * **New component: Menu**, [see docs](https://buefy.github.io/#/documentation/menu) (thanks @tsctao)
 * Add CDN install method and JSFiddle examples
+* Add boolean ``addons`` prop to field
 * Fix message and notification inside ``.media`` element
 * #171 Fix modal not taking full width on mobile
 * #182 Fix table not default sorting async data (thanks @jtommy)